Switched to using floats instead of unsigned ints for distances and more.
Now RGB5A3 data is not rounded by converting to RGBA8 before distances are checked.

+++ src/color_set.rs
@@ -16,7 +16,7 @@ impl<T: Color> ColorSet<T> {
 }
 
 impl<O: Color + Output> SimpleInput<ColorSet<O>> for ColorSet<Rgba8> {
-    fn distance_to(&self, other: &ColorSet<O>) -> u64 {
+    fn distance_to(&self, other: &ColorSet<O>) -> f64 {
         self.colors.iter().zip(other.colors.iter()).map(|(c1, c2)| c1.distance_to(c2)).sum()
     }
